AI_Employee_Vault/requirements.txt
watchdog>=3.0.0
psutil>=5.9.0
schedule>=1.2.0
python-dotenv>=1.0.0
PyYAML>=6.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.98.0
selenium>=4.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
ollama>=0.1.0